[133] Fixes an issue with the group_update action
Previously it would allow you to reassign the parent of a group, but
should you move a group from having a parent to no parent, it would not
remove the relationship with the existing one.  This now deletes the
member relationships where the current group is marked as a member of
a parent group.