[575] Forces a re-index of the datasets in a group when the name/slug changes
When a publisher has a slug change it loses its connection with all of the
datasets within that publisher.  This is only in WUI as the search index will
not have been updated.  To fix this we force a re-index of each dataset within
that group as we save it.

This is potentially problematic as it can take a while for the larger publishers
and it happens synchronously and it can take a little while.  I think at some point
we might be better placed to show an alert warning the user that the datasets are
being re-indexed and perform the actual indexing asynchronously.