[628] Fixed celery app so that overriding config actually works.
Previously celery_app was setting the config and then trying to overwrite it
based on the presence of an app:celery section in the config. This didn't work
as once you have told celery what its config options are it really doesn't want
to believe you when you tell it otherwise.